separated getEndpoint, updateAppState & executeRequest from ApiWorkbenchPresent
