+++ ts/ui/package.json
@@ -5,6 +5,7 @@
   "type": "module",
   "scripts": {
     "dev": "vite",
+    "typecheck": "tsc",
     "build": "tsc && vite build",
     "preview": "vite preview"
   },
