+++ ts/ui/src/navigation.tsx
@@ -3,7 +3,7 @@ import { RequireLogin } from "./components/RequireLogin";
 import { Messages } from "./pages/messages";
 import { Logout } from "./pages/logout";
 import { Login } from "./pages/login";
-import { ApiWorkbench } from "./pages/api-workbench";
+import { ApiWorkbench } from "./pages/api-view";
 
 export const ROUTES = {
   '/': () => <Landing />,
